% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funhist2df.R
\name{funhist2df}
\alias{funhist2df}
\title{Creates some plots and numeric summary}
\usage{
funhist2df(x)
}
\arguments{
\item{x}{vector}
}
\value{
the data frame with the summary
}
\description{
Creates some plots and numeric summary
}
